requirements.txt
Django==5.0.4
djangorestframework==3.15.1
pre-commit==3.7.0
PyJWT==2.8.0
python-dotenv==1.0.1
black==24.4.2
mysqlclient==2.2.4
drf-spectacular==0.27.2
drf-yasg==1.21.7